# source: Cache language config